add another check for mounted volume with "mount" output parsing
this is needed because on docker named volume all the content of /srv/data populated to the fresh volume automatically. This includes .not_mounted files as well and causes incorrect warning " changes will be lost on container restart". See https://github.com/umputun/tg-spam/issues/32#issuecomment-1869057449

In addition, I had to remove the VOLUME directive from the Dockerfile as well, because in this case volume is mounted inside the container regardless of the actual mount from the outside