Spam storage and web ui for detected spam (#51)
* Add spam logging to database

The main functionality of the spam logger has been expanded to not only log spam messages to a file, but also store them in a database. This facilitates efficient tracking and analysis of spam data. Furthermore, test cases have been added to ensure this new functionality works as expected.

* Add Detected Spam page to web UI

A Detected Spam page has been added to the web UI, where users can see a list of detected spam messages. Changes have been made to attach data DB to the server activation function and the timestamp is now localized. Fixed a bug in storage/detected_spam.go where the 'Checks' field was incorrectly attached to a non-existent 'entry' variable.

* Add DetectedSpamReader mock and implement detected spam HTML handler tests