it should address #246 and #245
The issue is the lack of mounted sample volumes,
and the default set of samples in this case is not
available for migration. To address it, the default
location of samples has changed to /srv/preset, and the
Dockerfile has been modified to keep them available.

This fix causes the re-import of preset samples every time
the container is recreated. It's not the most elegant
solution, but since this situation is temporary until
migrations are done, we can live with this for
the time being.